+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2024-09-18 -- v2.1.0
+- Query for (if necessary) and store patron's initial ptype code
+- Update to use black formatter
+
 ## 2024-06-17 -- v2.0.0
 - Query for (if necessary) and store patron's initial home library code
 - Use ET instead of EST when calculating all dates